// apps/web/data/gem-data.generated.ts
export interface GemData {
  name: string;
  castTime: number;
  critChance: number;
  damageEffectiveness: number;
  baseEffectiveness: number;
  incrementalEffectiveness: number;
  baseDamageMin: number;
  baseDamageMax: number;
  damageType: string;
  isSpell: boolean;
  tags: string;
  reqStr: number;
  reqDex: number;
  reqInt: number;
  isVaal: boolean;
  isSupport: boolean;
}